Extract data models to dedicated models.py module
- Move DatabricksRelationInfo (NamedTuple) and DatabricksConfig (dataclass) to models.py
- Remove 40 lines from impl.py by extracting core data structures
- Clean up unused imports: AdapterConfig, NamedTuple, TableFormat
- Improve code organization with dedicated models module
- Reduce impl.py from 1018 to 978 lines (additional 4% reduction)
- Total cumulative reduction: 76 lines (7.2% from original 1054 lines)
