+++ dbt/adapters/databricks/impl.py
@@ -8,7 +8,7 @@ from contextlib import contextmanager
 from dataclasses import dataclass
 from importlib import metadata
 from multiprocessing.context import SpawnContext
-from typing import TYPE_CHECKING, Any, ClassVar, Generic, NamedTuple, Optional, Union, cast
+from typing import TYPE_CHECKING, Any, ClassVar, Generic, Optional, Union, cast
 from uuid import uuid4
 
 from dbt_common.behavior_flags import BehaviorFlag
@@ -18,7 +18,7 @@ from dbt_common.utils import executor
 from dbt_common.utils.dict import AttrDict
 from packaging import version
 
-from dbt.adapters.base import AdapterConfig, PythonJobHelper
+from dbt.adapters.base import PythonJobHelper
 from dbt.adapters.base.impl import catch_as_completed, log_code_execution
 from dbt.adapters.base.meta import available
 from dbt.adapters.base.relation import BaseRelation
@@ -41,6 +41,7 @@ from dbt.adapters.databricks.column import DatabricksColumn
 from dbt.adapters.databricks.connections import DatabricksConnectionManager
 from dbt.adapters.databricks.global_state import GlobalState
 from dbt.adapters.databricks.handle import SqlUtils
+from dbt.adapters.databricks.models import DatabricksConfig, DatabricksRelationInfo
 from dbt.adapters.databricks.python_models.python_submissions import (
     AllPurposeClusterPythonJobHelper,
     JobClusterPythonJobHelper,
@@ -67,7 +68,6 @@ from dbt.adapters.databricks.relation_configs.materialized_view import (
 from dbt.adapters.databricks.relation_configs.streaming_table import (
     StreamingTableConfig,
 )
-from dbt.adapters.databricks.relation_configs.table_format import TableFormat
 from dbt.adapters.databricks.relation_configs.tblproperties import TblPropertiesConfig
 from dbt.adapters.databricks.relation_configs.view import ViewConfig
 from dbt.adapters.databricks.utils import get_first_row, handle_missing_objects, redact_credentials
@@ -79,46 +79,6 @@ if TYPE_CHECKING:
 
 dbt_version = metadata.version("dbt-core")
 SUPPORT_MICROBATCH = version.parse(dbt_version) >= version.parse("1.9.0b1")
-
-
-class DatabricksRelationInfo(NamedTuple):
-    table_name: str
-    table_type: str
-    file_format: Optional[str]
-    table_owner: Optional[str]
-    databricks_table_type: Optional[str]
-
-
-@dataclass
-class DatabricksConfig(AdapterConfig):
-    file_format: str = "delta"
-    table_format: str = TableFormat.DEFAULT
-    location_root: Optional[str] = None
-    include_full_name_in_path: bool = False
-    partition_by: Optional[Union[list[str], str]] = None
-    clustered_by: Optional[Union[list[str], str]] = None
-    liquid_clustered_by: Optional[Union[list[str], str]] = None
-    auto_liquid_cluster: Optional[bool] = None
-    buckets: Optional[int] = None
-    options: Optional[dict[str, str]] = None
-    merge_update_columns: Optional[str] = None
-    merge_exclude_columns: Optional[str] = None
-    databricks_tags: Optional[dict[str, str]] = None
-    tblproperties: Optional[dict[str, str]] = None
-    zorder: Optional[Union[list[str], str]] = None
-    unique_tmp_table_suffix: bool = False
-    skip_non_matched_step: Optional[bool] = None
-    skip_matched_step: Optional[bool] = None
-    matched_condition: Optional[str] = None
-    not_matched_condition: Optional[str] = None
-    not_matched_by_source_action: Optional[str] = None
-    not_matched_by_source_condition: Optional[str] = None
-    target_alias: Optional[str] = None
-    source_alias: Optional[str] = None
-    merge_with_schema_evolution: Optional[bool] = None
-    use_safer_relation_operations: Optional[bool] = None
-    incremental_apply_config_changes: Optional[bool] = None
-    view_update_via_alter: Optional[bool] = None
 
 
 def get_identifier_list_string(table_names: set[str]) -> str:
